requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
langchain>=1.0.0
langchain-classic>=0.3.14
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-google-genai>=3.2.0
langchain-xai>=1.0.0
langgraph>=1.0.0
google-api-core>=2.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
openai>=1.109.1,<3.0.0
httpx>=0.28.0
datadog-api-client>=2.32.0
motor>=3.7.0
python-jose[cryptography]>=3.3.0
pytest>=8.3.4
pytest-asyncio>=0.23.0
